Add files via upload
This case study aims to give an idea of applying EDA in a real business scenario. In this case study, applied EDA techniques to develop a basic understanding of risk analytics in banking and financial services and understand how data is used to minimise the risk of losing money while lending to customers.